ref(quotas): Split crate types into submodules (#600)
Refactors `relay-quotas` without any functional changes and adds a few
tests.
