+++ relay-quotas/Cargo.toml
@@ -11,7 +11,7 @@ publish = false
 
 [features]
 default = []
-rate-limiter = [
+redis = [
     "failure",
     "log",
     "relay-redis/impl",
